ingestion-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
confluent-kafka
python-dotenv